Renamed mekhq.plugin.* to mekhq.module.*
module is a better name because it includes support for built-in
services as well as plugins.
